requirements.txt
streamlit
svgwrite
python-dotenv
openai
supabase
